refactor(mvc): remove view dependency from Dealer and delegate turn control to TurnManager
- Moved askPlayerDraw and playDealerTurn logic out of Dealer
- Dealer no longer depends on InputView or OutputView
- Added TurnManager.handleDealerTurn to coordinate dealer drawing with view updates
- Passed "Dealer" name explicitly to Participant to preserve display needs
