+++ src/main/kotlin/blackjack/model/Dealer.kt
@@ -1,9 +1,6 @@
 package blackjack.model
 
-import blackjack.view.InputView
-import blackjack.view.OutputView
-
-open class Dealer : Participant("Dealer") {
+open class Dealer : Participant(name = "Dealer") {
     private val deck = Deck()
 
     fun giveTwoCardsTo(participants: List<Participant>) {
@@ -18,29 +15,8 @@ open class Dealer : Participant("Dealer") {
         participant.addCard(card)
     }
 
-    fun askPlayerDraw(player: Player) {
-        while (true) {
-            val wantsToDraw = InputView.askPlayerWantsToDraw(player.name)
-            if (wantsToDraw) {
-                giveCardTo(player)
-                OutputView.displayPlayerHand(player)
-                if (player.getScore() >= 21) break
-            } else {
-                break
-            }
-        }
-    }
-
-    fun playDealerTurn() {
-        while (shouldDraw()) {
-            OutputView.displayDealerDraw()
-            giveCardTo(this)
-            OutputView.displayDealerStatus(this)
-        }
-    }
-
     fun shouldDraw(): Boolean {
-        return (this.getScore()) <= 16
+        return getScore() <= 16
     }
 
     fun judge(player: Player): GameResult {
